Исправлена ошибка с сортировкой, switch-case по дефолту выдает RuntimeException с неправильной командой сортировки

+++ sandbox.controller/src/kz/greetgo/sandbox/controller/model/ClientRecord.java
@@ -10,7 +10,7 @@ public class ClientRecord {
 
   //Khamit puchemu long? - 1
 
-  public long totalAccountBalance;
-  public long maxAccountBalance;
-  public long minAccountBalance;
+  public float totalAccountBalance;
+  public float maxAccountBalance;
+  public float minAccountBalance;
 }
